requirements.txt
# Python 3
attrs
click
git+https://github.com/jwodder/headerparser
requests
